SGX-BIN: Fix incorrect tiler page addresses on tiler2d allocations
For some reason, the items in the tiler2d scatterlist has a different
stride value, then what omap_tiler_pages does. the scatterlist
operates with 0x1000 strides, while a NV12 allocation will have 0x4000
on the Y plane, and 0x8000 on the UV plane.

This seems to relate to section 15.2.3.6.1.8.2 on TRM (vZ)s stride values.

This was inspired by how this was handled in SGX DDK 1.9.

Change-Id: I72fce274576d9b32e7e1d892cbcb168338032a91
Signed-off-by: D. Andrei Măceș <andrei@unlegacy-android.org>